fix: make SendGrid secret conditional in Terraform
Azure Container Apps doesn't allow empty secrets, so we need to only
create the sendgrid-api-key secret when it has a value.

Changes:
- Use dynamic block to conditionally create sendgrid-api-key secret
- Remove SENDGRID_API_KEY env var (not needed for noop provider)
- Secret only created when var.sendgrid_api_key != ""

This fixes Terraform failure:
"Container app secret(s) with name(s) 'sendgrid-api-key' are invalid:
 value or keyVaultUrl and identity should be provided."

When using noop or smtp providers, we don't need SendGrid credentials.
When switching to SendGrid later, update the workflow to provide the key.

Signed-off-by: Jose Blanquicet <josebl@microsoft.com>